+++ tests/bloom_multi_thread.c
@@ -1,7 +1,7 @@
 
 #include <stdlib.h>
 #include <stdio.h>
-#include "../bloom.h"
+#include "../src/bloom.h"
 #include "../timing-c/timing.h" /* URL: https://github.com/barrust/timing-c */
 #include <omp.h>
